+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [Unreleased]
 ### Added
 - Added types to getCommunityInfo.
+- Added getCommunityReminder method.
 
 ### Changed
 - Refactored a bunch of interfaces.